removing old front js and reference to facia
