+++ common/app/assets/javascripts/bootstraps/front.js
@@ -1,84 +1,107 @@
 define([
-    //Common libraries
-    "common",
-    "bonzo",
-    "domReady",
-    //Modules
-    "modules/trailblocktoggle",
-    "modules/trailblock-show-more",
-    "modules/footballfixtures",
-    "modules/cricket"
+    // Common libraries
+    'common',
+    'bonzo',
+    // Modules
+    'modules/detect',
+    'modules/facia/popular',
+    'modules/facia/items-show-more',
+    'modules/facia/collection-display-toggle',
+    'modules/footballfixtures',
+    'modules/facia/image-upgrade',
+    'modules/cricket'
 ], function (
     common,
     bonzo,
-    domReady,
-    TrailblockToggle,
-    TrailblockShowMore,
+    detect,
+    popular,
+    ItemsShowMore,
+    CollectionDisplayToggle,
     FootballFixtures,
-    Cricket
-) {
+    ImageUpgrade,
+    cricket
+    ) {
 
     var modules = {
 
         showCricket: function(){
             common.mediator.on('page:front:ready', function(config, context) {
-                Cricket.cricketTrail(config, context);
+                cricket.cricketTrail(config, context);
             });
         },
 
-        showTrailblockToggles: function () {
-            var tt = new TrailblockToggle();
+        showPopular: function () {
             common.mediator.on('page:front:ready', function(config, context) {
-                tt.go(config, context);
+                popular.render(config);
             });
         },
 
-        showTrailblockShowMore: function () {
-            var trailblockShowMore = new TrailblockShowMore();
+        showItemsShowMore: function () {
             common.mediator.on('page:front:ready', function(config, context) {
-                trailblockShowMore.init(context);
+                common.$g('.js-items--show-more', context).each(function(items) {
+                    new ItemsShowMore(items)
+                        .addShowMore();
+                });
             });
         },
 
         showFootballFixtures: function(path) {
             common.mediator.on('page:front:ready', function(config, context) {
-                if(config.page.edition === "UK") {
-
-                    var opts,
-                        table;
-
-                    switch(config.page.pageId) {
-                        case "" :
-                            // Network Front
-                            opts = {
-                                prependTo: context.querySelector('.zone-sport ul > li'),
-                                competitions: ['500', '510', '100', '400'],
-                                contextual: false,
-                                expandable: true,
-                                numVisible: 3
-                            };
-                            break;
-                        case "sport" :
-                            // Sport Front
-                            // don't want to put it in the masthead trailblock
-                            var trailblocks = [].filter.call(context.querySelectorAll('.trailblock'), function(trailblock) {
-                                return bonzo(trailblock).hasClass('trailblock--masthead') === false;
-                            });
-                            opts = {
-                                prependTo: (trailblocks.length) ? trailblocks[0].querySelector('ul > li') : null,
-                                competitions: ['500', '510', '100', '400'],
-                                contextual: false,
-                                expandable: true,
-                                numVisible: 5
-                            };
-                            break;
-                    }
-
-                    if(opts && !bonzo(opts.prependTo).hasClass('footballfixtures-loaded')) {
-                        bonzo(opts.prependTo).addClass('footballfixtures-loaded');
-                        table = new FootballFixtures(opts).init();
-                    }
+                if (config.page.edition === 'UK' && (config.page.pageId === "" || config.page.pageId === "sport")) {
+                    // wrap the return sports stats component in an 'item'
+                    var $statsItem = bonzo(bonzo.create('<li class="item item--sport-stats"></li>')),
+                        section = config.page.pageId === "" ? 'sport' : 'news',
+                        numVisible = config.page.pageId === "" ? 3 : 5;
+                    common.mediator.on('modules:footballfixtures:render', function() {
+                        var container = common.$g('.collection--' + section, context)
+                            .first()[0];
+                        if (container) {
+                            // toggle class
+                            common.$g('.items', container)
+                                .removeClass('items--without-sport-stats')
+                                .addClass('items--with-sport-stats');
+                            // add it after the first item
+                            common.$g('.item:first-child', container)
+                                .after($statsItem);
+                            // now hide one of the shown ones (but not on mobile)
+                            if (detect.getBreakpoint() !== 'mobile') {
+                                common.$g('.item.u-h', container)
+                                    .first()
+                                    .previous()
+                                    .addClass('u-h');
+                            }
+                        }
+                    });
+                    new FootballFixtures({
+                        prependTo: $statsItem,
+                        attachMethod: 'append',
+                        competitions: ['500', '510', '100', '400'],
+                        contextual: false,
+                        expandable: true,
+                        numVisible: numVisible
+                    }).init();
                 }
+            });
+        },
+
+        showCollectionDisplayToggle: function () {
+            common.mediator.on('page:front:ready', function(config, context) {
+                common.$g('.js-collection--display-toggle', context).each(function(collection) {
+                    new CollectionDisplayToggle(collection, config)
+                        .addToggle();
+                });
+            });
+        },
+
+        upgradeImages: function () {
+            common.mediator.on('page:front:ready', function(config, context) {
+                common.$g('.collection', context).each(function(collection) {
+                    var isContainer = (bonzo(collection).attr('data-collection-type') === 'container');
+                    common.$g('.item', collection).each(function(item, index) {
+                        new ImageUpgrade(item, isContainer && (index === 0))
+                            .upgrade();
+                    });
+                });
             });
         }
 
@@ -87,10 +110,11 @@ define([
     var ready = function (config, context) {
         if (!this.initialised) {
             this.initialised = true;
-            modules.showTrailblockToggles();
-            modules.showTrailblockShowMore();
+            modules.showPopular();
+            modules.showItemsShowMore();
             modules.showFootballFixtures();
-            modules.showCricket();
+            modules.showCollectionDisplayToggle();
+            modules.upgradeImages();
         }
         common.mediator.emit("page:front:ready", config, context);
     };
