+++ common/app/views/fragments/javaScriptConfig.scala.html
@@ -13,7 +13,6 @@
         "edition": "@Edition(request).id",
         "ajaxUrl": "@Configuration.ajax.url",
         "isDev": @play.Play.isDev(),
-        "isFacia" : @IsFacia(request),
         "oasHost": "oas.theguardian.com",
         "oasUrl": "http://oas.theguardian.com/RealMedia/ads/",
         "oasSiteIdHost": "www.theguardian.com",
